FERandom to Weapons
Updated the basic Weapon randomization to use the FERandom module, so the chosen distribution/memory effects apply to those statistics as well.  This should have the effect of, for example, when memory = true it should hopefully keep there from being a streak of really low power or low hit weapons.