Custom Class Growths Option
Added the ability to load custom class growths from a .json file for the GBA games.  Untested for 6&7, and additional classes will need to be added for those games.  Integration into the UI seems complete and the feature seems to be working as expected.  Possibly need to playtest as enemies might have become quite a bit tougher.